New updates made
On HTML file- added ids to all the objects that will be affected on the script file, that did not have ids yet, such as humidity or icon.

On script file- made sure that description, the time, humidity, wind speed, and the icon changed depending on the conditions and the city searched.
Also added another function to make sure that the initial page has a city already searched so that those areas of temperature are not left blank, like the HTML file is.